+++ src/main/java/frc/robot/components/FileLogging.java
@@ -4,44 +4,24 @@ import java.io.*;
 
 public class FileLogging {
     private String filename, out, dir;
-    private final String DEFAULT_PATH = "/1559data/";
 
-    /**
-     * A file mannagment class for creating and reading files
-     */
     public FileLogging() {
         out = "";
         filename = "";
         dir = "";
     }
 
-    /**
-     * Adds data to be written
-     * 
-     * @param data New data to be written
-     */
-    public void addData(String data) {
+    public void periodic(String data) {
         out += data;
     }
 
-    /**
-     * Sets the directory that the file will be written in and creates the direcory
-     * if it doesn't exist (Doesn't create the directory on roborios)
-     * 
-     * @param dir The directory to be set
-     */
     public void setDirectory(String dir) {
         if (new File(dir).mkdir()) {
             System.out.println("Created directory");
         }
-        this.dir = DEFAULT_PATH + dir;
+        this.dir = "/1559data/" + dir;
     }
 
-    /**
-     * Creates a file with the given name
-     * 
-     * @param name Name of the file to be written
-     */
     public void createfile(String name) {
         if (name.equals("")) {
             name = dir + "Default_file.txt";
@@ -71,19 +51,11 @@ public class FileLogging {
         }
     }
 
-    /**
-     * Writes the data stored using the addData() method
-     */
     public void write() {
         write(out);
         out = "";
     }
 
-    /**
-     * Writes data to the last file created
-     * 
-     * @param data Data to be written
-     */
     public void write(String data) {
         try {
             FileWriter myWriter = new FileWriter(dir + filename);
@@ -99,42 +71,10 @@ public class FileLogging {
         }
     }
 
-    /**
-     * Writes data to the designated file
-     * 
-     * @param filename Name of the file to be written
-     * @param data     Data to be written
-     */
-    public void write(String filename, String data) {
-        try {
-            FileWriter myWriter = new FileWriter(dir + filename);
-
-            myWriter.write(data);
-            myWriter.close();
-            // System.out.println("Successfully wrote to the file.");
-        }
-
-        catch (IOException e) {
-            System.out.println("An error occurred.");
-            // e.printStackTrace();
-        }
-    }
-
-    /**
-     * Reads the last file created
-     * 
-     * @return The contents os the file
-     */
     public String readFile() {
         return readFile(filename);
     }
 
-    /**
-     * Reads the file with the given name
-     * 
-     * @param fileName File to be read
-     * @return The contents of the file
-     */
     public String readFile(String fileName) {
         File file = new File(dir + fileName);
         String fileContent = "";
@@ -152,5 +92,19 @@ public class FileLogging {
 
         return fileContent;
 
+    }
+
+    public double interpolate(double counter, double[] value) {
+        int intCounter = (int) counter;
+        double percent = (counter - intCounter);
+
+        if (intCounter < value.length - 1) {
+            double interpolatedValue = (value[intCounter] + (percent * (value[intCounter + 1] - value[intCounter])));
+            return interpolatedValue;
+        }
+
+        else {
+            return value[value.length - 1];
+        }
     }
 }
